requirements.txt
time
cv2
openai
python-dotenv
wave
pyaudio
Pillow
numpy
typing-extensions
simpleaudio
warnings